feat: alteração das extensões .js para .jsx e mudança de css para module.css
