+++ src/App.jsx
@@ -1,16 +1,16 @@
 import React from "react";
-import {PrimeReactProvider} from "primereact/api";
+import { PrimeReactProvider } from "primereact/api";
 import "primereact/resources/themes/lara-light-blue/theme.css";
 import "primereact/resources/primereact.min.css";
 import "primeicons/primeicons.css";
 import AppRoutes from "./AppRoutes";
 
-import './App.css'
+import './App.module.css';
 
 const App = () => {
     return (
         <PrimeReactProvider>
-            <AppRoutes/> {/* Renderiza as rotas corretamente */}
+            <AppRoutes />
         </PrimeReactProvider>
     );
 }
