+++ src/App.module.css
@@ -1,4 +1,4 @@
-body {
+:global(body) {
     margin: 0;
     padding: 0;
 }
