+++ src/pages/register/Register.module.css
@@ -1,4 +1,4 @@
-#container-register {
+.container {
     background-color: #677FFA;
     width: 100vw;
     height: 100vh;
@@ -11,7 +11,7 @@
     box-sizing: border-box;
 }
 
-#custom-card {
+.card {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -24,39 +24,39 @@
     text-align: center;
 }
 
-#register-text {
+.title {
     font-size: 1.5rem;
     font-weight: bold;
     margin: 0;
     padding: 0;
 }
 
-#register-description {
+.description {
     font-size: 1rem;
     font-weight: normal;
     color: #8C95A4;
     margin-bottom: 2.1875rem;
 }
 
-form {
+.form {
     display: flex;
     flex-direction: column;
     gap: 1rem;
     align-items: center;
 }
 
-#checkbox-container {
+.checkboxContainer {
     display: flex;
     align-items: center;
     gap: 0.5rem;
     margin-top: 1rem;
 }
 
-#checkbox {
+.checkbox {
     margin: 0;
 }
 
-#register-button {
+.registerButton {
     background-color: #677FFA;
     width: 80%;
     margin-top: 1rem;
@@ -67,14 +67,13 @@ form {
     cursor: pointer;
 }
 
-#register-button:hover {
+.registerButton:hover {
     background-color: #5567d9;
 }
 
-#log-in-button {
+.loginRedirect {
     margin-bottom: 0;
     text-align: center;
     color: #677FFA;
     cursor: pointer;
 }
-
