+++ src/pages/requestNewPassword/RequestNewPassword.module.css
@@ -1,4 +1,4 @@
-#container-request {
+.container {
     background-color: #677FFA;
     width: 100vw;
     height: 100vh;
@@ -11,7 +11,7 @@
     box-sizing: border-box;
 }
 
-#custom-card {
+.card {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -24,28 +24,28 @@
     text-align: center;
 }
 
-#request-text {
+.title {
     font-size: 1.5rem;
     font-weight: bold;
     margin: 0;
     padding: 0;
 }
 
-#request-description {
+.description {
     font-size: 1rem;
     font-weight: normal;
     color: #8C95A4;
     margin-bottom: 2.1875rem;
 }
 
-form {
+.form {
     display: flex;
     flex-direction: column;
     gap: 1rem;
     align-items: center;
 }
 
-#request-button {
+.requestButton {
     background-color: #677FFA;
     width: 80%;
     margin-top: 1rem;
@@ -56,7 +56,6 @@ form {
     cursor: pointer;
 }
 
-#request-button:hover {
+.requestButton:hover {
     background-color: #5567d9;
 }
-
